DAO: Fixes bug causing fk constrait failure
The bug caused a SQLiteException with error code 19 (a foreign key
constraint was violated) when adding accounts, specifically when linking
account ancestors to descendants.

It was caused by the descendants being added before they had been
inserted into the database as accounts. As the foreign key constraints
on AncestorTo dictates that the descendant parameter is a key on an
account name, and no account existed with that name, SQLite threw a
foreign constraints error.

The solution involved refactoring the code that inserts accounts and
links them by ancestor/descendant, so that all accounts in an account
string are added before the linking begins.