DAO: Fixes misconceived insert account solution
As an example, if you specified "A.B.C" as an account string before, the
accounts "A", "B" and "C" would be added. "C" would be set as a
descendant of "A" and "B" and "B" and "C" would be set as a descendant
of "A".

This was a misconcieved solution because its not how we want the
accounts to be stored. If you have "Assets.Checking" you don't want the
"Checking" account to be named just that, it should be named
"Assets.Checking".

This is now fixed so the account string "A.B.C" gets inserted as
accounts "A", "A.B" and "A.B.C". "A.B.C" get stored as a descendant of
"A.B" and "A.B.C" and "A.B" stored as a descendant of "A".