Fixes post exception not caught bug
The bug was that exceptions thrown in POST controller handler methods
where not caught by the developer exception page middleware and was
caused by the controller post handlers being async methods and at the
same time returning `void`. In order for the developer exception
middleware to be able to catch async exceptions, the return type must
be `Task`.