Adds custom model binder for account
Change was made because I needed a way to deserialize the form data
fields 'debitAccount' and 'creditAccount' to entities in the database.

The databinding works by looking at the form value for 'debitAccount'
and 'creditAccount' fields and searching in the database for account
entities that have the same name.

To make this work, I also needed to change the POST content type from
JSON to x-www-form-urlencoded since custom model binders for
properties only works with form data.